+++ src/intersection.c
@@ -12,7 +12,7 @@ Intersection NewIntersection(Shape* s, Ray r) {
     i.count = 0;
     i.shape_ptr = s;
     i.ray = r;
-    memset(i.ray_times, 0, sizeof(float) * MAX_NUMBER_INTERSECTIONS);
+    memset(i.ray_times, 0, sizeof(double) * MAX_NUMBER_INTERSECTIONS);
     return i;
 }
 
@@ -41,10 +41,10 @@ Intersection IntersectSphere(Shape* s, Ray r) {
     r = RayTransform(r, s->inverse_transform);
 
     Tuple3 sphere_to_ray = TupleSubtract(r.origin, NewPnt3(0, 0, 0));
-    float a = TupleDotProduct(r.direction, r.direction);
-    float b = 2 * TupleDotProduct(r.direction, sphere_to_ray);
-    float c = TupleDotProduct(sphere_to_ray, sphere_to_ray) - 1;
-    float discriminant = (b * b) - 4 * a * c;
+    double a = TupleDotProduct(r.direction, r.direction);
+    double b = 2 * TupleDotProduct(r.direction, sphere_to_ray);
+    double c = TupleDotProduct(sphere_to_ray, sphere_to_ray) - 1;
+    double discriminant = (b * b) - 4 * a * c;
     
     if (discriminant < 0) {
         return result;
@@ -57,15 +57,15 @@ Intersection IntersectSphere(Shape* s, Ray r) {
     } else {
         result.count = 2;
 
-        float a_2 = a * 2;
-        float d_sqrt = sqrtf(discriminant);
+        double a_2 = a * 2;
+        double d_sqrt = sqrt(discriminant);
 
         result.ray_times[0] = (-b - d_sqrt) / a_2;
         result.ray_times[1] = (-b + d_sqrt) / a_2;
     }
 
-    float max = fmaxf(result.ray_times[0], result.ray_times[1]);
-    float min = fminf(result.ray_times[0], result.ray_times[1]);
+    double max = fmax(result.ray_times[0], result.ray_times[1]);
+    double min = fmin(result.ray_times[0], result.ray_times[1]);
     result.ray_times[0] = min;
     result.ray_times[1] = max;
